# perf(remix-dev/vite): disable `server.preTransformRequests` for child compiler