--- a/.changeset/short-phones-admire.md
+++ b/.changeset/short-phones-admire.md
@@ -1,0 +1,5 @@
+---
+"@remix-run/dev": patch
+---
+
+Improve performance of Vite child compiler by disabling `server.preTransformRequests`