@pcattori 2023-11-25
Not sure if this API is available in Vite 5. Looking at [this repo](https://github.com/bluwy/vite-plugin-warmup?tab=readme-ov-file#vite-plugin-warmup), seems like [`server.warmup`](https://vitejs.dev/guide/performance.html#warm-up-frequently-used-files) is now the intended API. But I haven't looked into it deeply, so not sure.