@hi-ogawa 2023-11-26
I think what `server.warmup` does it to provide a more aggressive way to trigger `warmupRequest` at the launch of the server.
https://github.com/vitejs/vite/blob/227d56d37fbfcd1af4b5d93182770b4e650511ee/packages/vite/src/node/server/index.ts#L776-L782

The above code snippet I copied from `plugins/importAnalysis.ts` is still the default heuristic/optimization done by Vite internally to trigger `warmupRequest` when Vite's import analysis finds static import of local files since it's likely for users to request transform of such dependency right after that. But for the case of child compiler, this heuristics is "false-positive" in a sense and I think remix can disable it to reduce unnecessary work.
It seems this `preTransformRequests` option initially came from similar needs on vitest https://github.com/vitejs/vite/pull/6309 https://github.com/vitest-dev/vitest/issues/229